Move factory to a better file structure
