+++ lib/tv/sony.py
@@ -1,0 +1,6 @@
+
+
+class TvSony: # implements TV contract
+
+    def getName(self):
+        return "Sony TV here"
